+++ crates/nostr-sdk/examples/blacklist.rs
@@ -27,7 +27,10 @@ async fn main() -> Result<()> {
         .authors([muted_public_key, public_key])
         .kind(Kind::Metadata);
     let events = client
-        .get_events_of(vec![filter], Some(Duration::from_secs(10)))
+        .get_events_of(
+            vec![filter],
+            EventSource::relays(Some(Duration::from_secs(10))),
+        )
         .await?;
     println!("Received {} events.", events.len());