+++ crates/nostr-sdk/examples/nip65.rs
@@ -19,7 +19,10 @@ async fn main() -> Result<()> {
 
     let filter = Filter::new().author(public_key).kind(Kind::RelayList);
     let events: Vec<Event> = client
-        .get_events_of(vec![filter], Some(Duration::from_secs(10)))
+        .get_events_of(
+            vec![filter],
+            EventSource::relays(Some(Duration::from_secs(10))),
+        )
         .await?;
     let event = events.first().unwrap();
     println!("Found relay list metadata:");